wip clearing up utils.py
