+++ src/spatialdata/_io/_utils.py
@@ -1,13 +1,24 @@
+from __future__ import annotations
+
+import filecmp
 import logging
+import os.path
+import re
+import tempfile
 from collections.abc import Generator, Mapping
 from contextlib import contextmanager
+from functools import singledispatch
 from typing import Any, Optional, Union
 
 import zarr
+from dask.dataframe.core import DataFrame as DaskDataFrame
 from multiscale_spatial_image import MultiscaleSpatialImage
 from ome_zarr.format import Format
 from ome_zarr.writer import _get_valid_axes
+from spatial_image import SpatialImage
+from xarray import DataArray
 
+from spatialdata import SpatialData
 from spatialdata._core.core_utils import (
     MappingToCoordinateSystem_t,
     ValidAxis_t,
@@ -18,6 +29,7 @@ from spatialdata._core.transformations import (
     BaseTransformation,
     _get_current_output_axes,
 )
+from spatialdata.utils import iterate_pyramid_levels
 
 
 # suppress logger debug from ome_zarr with context manager
@@ -123,3 +135,106 @@ def _iter_multiscale(
         return [getattr(data[i][name], attr) for i in data.keys()]
     else:
         return [data[i][name] for i in data.keys()]
+
+
+class dircmp(filecmp.dircmp):  # type: ignore[type-arg]
+    """
+    Compare the content of dir1 and dir2. In contrast with filecmp.dircmp, this
+    subclass compares the content of files with the same path.
+    """
+
+    def phase3(self) -> None:
+        """
+        Find out differences between common files.
+        Ensure we are using content comparison with shallow=False.
+        """
+        fcomp = filecmp.cmpfiles(self.left, self.right, self.common_files, shallow=False)
+        self.same_files, self.diff_files, self.funny_files = fcomp
+
+
+def _are_directories_identical(
+    dir1: Any,
+    dir2: Any,
+    exclude_regexp: Optional[str] = None,
+    _root_dir1: Optional[str] = None,
+    _root_dir2: Optional[str] = None,
+) -> bool:
+    """
+    Compare two directory trees content.
+    Return False if they differ, True is they are the same.
+    """
+    if _root_dir1 is None:
+        _root_dir1 = dir1
+    if _root_dir2 is None:
+        _root_dir2 = dir2
+    if exclude_regexp is not None:
+        if re.match(rf"{_root_dir1}/" + exclude_regexp, str(dir1)) or re.match(
+            rf"{_root_dir2}/" + exclude_regexp, str(dir2)
+        ):
+            return True
+
+    compared = dircmp(dir1, dir2)
+    if compared.left_only or compared.right_only or compared.diff_files or compared.funny_files:
+        return False
+    for subdir in compared.common_dirs:
+        if not _are_directories_identical(
+            os.path.join(dir1, subdir),
+            os.path.join(dir2, subdir),
+            exclude_regexp=exclude_regexp,
+            _root_dir1=_root_dir1,
+            _root_dir2=_root_dir2,
+        ):
+            return False
+    return True
+
+
+def _compare_sdata_on_disk(a: SpatialData, b: SpatialData) -> bool:
+    from spatialdata import SpatialData
+
+    if not isinstance(a, SpatialData) or not isinstance(b, SpatialData):
+        return False
+    # TODO: if the sdata object is backed on disk, don't create a new zarr file
+    with tempfile.TemporaryDirectory() as tmpdir:
+        a.write(os.path.join(tmpdir, "a.zarr"))
+        b.write(os.path.join(tmpdir, "b.zarr"))
+        return _are_directories_identical(os.path.join(tmpdir, "a.zarr"), os.path.join(tmpdir, "b.zarr"))
+
+
+def _get_backing_files_raster(raster: DataArray) -> list[str]:
+    files = []
+    for k, v in raster.data.dask.layers.items():
+        if k.startswith("original-from-zarr-"):
+            mapping = v.mapping[k]
+            path = mapping.store.path
+            files.append(os.path.realpath(path))
+    return files
+
+
+@singledispatch
+def get_backing_files(element: Union[SpatialImage, MultiscaleSpatialImage, DaskDataFrame]) -> list[str]:
+    raise TypeError(f"Unsupported type: {type(element)}")
+
+
+@get_backing_files.register(SpatialImage)
+def _(element: SpatialImage) -> list[str]:
+    return _get_backing_files_raster(element)
+
+
+@get_backing_files.register(MultiscaleSpatialImage)
+def _(element: MultiscaleSpatialImage) -> list[str]:
+    xdata0 = next(iter(iterate_pyramid_levels(element)))
+    return _get_backing_files_raster(xdata0)
+
+
+@get_backing_files.register(DaskDataFrame)
+def _(element: DaskDataFrame) -> list[str]:
+    files = []
+    layers = element.dask.layers
+    for k, v in layers.items():
+        if k.startswith("read-parquet-"):
+            t = v.creation_info["args"]
+            assert isinstance(t, tuple)
+            assert len(t) == 1
+            parquet_file = t[0]
+            files.append(os.path.realpath(parquet_file))
+    return files
