+++ tests/test_utils.py
@@ -14,11 +14,8 @@ from xarray import DataArray
 from spatialdata import read_zarr
 from spatialdata._core.models import get_schema
 from spatialdata._core.transformations import Affine
-from spatialdata.utils import (
-    get_backing_files,
-    multiscale_spatial_image_from_data_tree,
-    unpad_raster,
-)
+from spatialdata._io._utils import get_backing_files
+from spatialdata.utils import multiscale_spatial_image_from_data_tree, unpad_raster
 
 
 def _pad_raster(data: DataArray, axes: tuple[str, ...]) -> DataArray:
